removed ua from ga
